Fixed Authentication for new Roles
Changed the role from Company and Association to Dono and Beneficiary
Introduced field *verfied* in User model. It can be used to check if the account got the permission requested from admin, is required to be active to open campaign or to receive donations.
A *verfied = false* will not block the login.